fix(user-service): fix bigint telegramId transformer and jwt expiredAt overflow
- TelegramId transformer: convert string from pg bigint to number
- JwtResponseDto: change expiredAt from Int to Float (Date.now() exceeds 32-bit)

+++ libs/contracts/src/dto/user-service/jwt-response.dto.ts
@@ -1,10 +1,10 @@
-import { Field, Int, ObjectType } from '@nestjs/graphql'
+import { Field, Float, ObjectType } from '@nestjs/graphql'
 
 @ObjectType()
 export class JwtResponseDto {
     @Field()
     accessToken!: string
 
-    @Field(() => Int)
+    @Field(() => Float)
     expiredAt!: number
 }
